+++ src/main/java/com/zb/cinema/admin/model/request/InputTheater.java
@@ -1,0 +1,20 @@
+package com.zb.cinema.admin.model.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class InputTheater {
+
+    private String area;
+    private String city;
+    private String name;
+}
