--- src/main/java/com/zb/cinema/movie/model/InputDates.java
@@ -1,16 +1,0 @@
-package com.zb.cinema.movie.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class InputDates {
-
-    private String startDt;
-    private String endDt;
-}
