+++ src/main/java/com/zb/cinema/movie/model/request/InputDate.java
@@ -1,17 +1,18 @@
-package com.zb.cinema.admin.model;
+package com.zb.cinema.movie.model.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class InputTheater {
+public class InputDate {
 
-    private String area;
-    private String city;
-    private String name;
+    private String date;
 }
