+++ src/main/java/com/zb/cinema/movie/model/request/InputMovieCode.java
@@ -1,15 +1,18 @@
-package com.zb.cinema.movie.model;
+package com.zb.cinema.movie.model.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class InputDate {
+public class InputMovieCode {
 
-    private String date;
+    private Long movieCode;
 }
