+++ src/main/java/com/zb/cinema/movie/model/response/ResponseMessage.java
@@ -1,12 +1,15 @@
-package com.zb.cinema.movie.model;
+package com.zb.cinema.movie.model.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.http.HttpStatus;
 
-@Data
+@Getter
+@Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
