+++ src/main/java/com/zb/cinema/movie/model/response/ResponseMessageHeader.java
@@ -1,11 +1,14 @@
-package com.zb.cinema.movie.model;
+package com.zb.cinema.movie.model.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
